move RedactingWriter from pkg/flare to pkg/util/scrubber
